refactor(dashboard): move Lens by-reference drilldown helpers to lenscommon
